requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
Faker==37.4.2
pillow==11.1.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
